small changes and added the planned project structure
